Initial commit
I didn't behave nicely and committed while working on this. The
functionality I envisioned is there. It could work and look a little
better (and faster) but it definitely works. To test it: put these 3
files in the root of your local stacks-website. I need to find a way to
integrate things more nicely at some point.